Move Utils.GetTestProjectIdFromEnvironment and Utils.IsWindows
The test ID generation will come separately.

Initially I wanted to use fixtures as we do elsewhere, but that
would have been a more invasive change for very little benefit.
